refactor: reorganize database files, create Database Module, and
