+++ src/infra/http/controllers/create-question.controller.ts
@@ -4,7 +4,7 @@ import { JwtAuthGuard } from '@/infra/auth/jwt-auth.guard'
 import { UserPayload } from '@/infra/auth/jwt.strategy'
 import { z } from 'zod'
 import { ZodValidationPipe } from '@/infra/http/pipes/zod-validation-pipe'
-import { PrismaService } from '@/infra/prisma/prisma.service'
+import { PrismaService } from '@/infra/http/database/prisma/prisma.service'
 
 const createQuestionBodySchema = z.object({
   title: z.string(),
