+++ src/infra/http/database/database.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common'
+import { PrismaService } from './prisma/prisma.service'
+
+@Module({
+  providers: [PrismaService],
+  exports: [PrismaService],
+})
+export class DatabaseModule {}
