+++ src/infra/http/http.module.ts
@@ -3,7 +3,7 @@ import { CreateAccountController } from './controllers/create-account.controller
 import { AuthenticateController } from './controllers/authenticate.controller'
 import { CreateQuestionController } from './controllers/create-question.controller'
 import { FetchRecentQuestionsController } from './controllers/fetch-recent-questions.controller'
-import { PrismaService } from '../prisma/prisma.service'
+import { PrismaService } from './database/prisma/prisma.service'
 
 @Module({
   controllers: [
